Fix the Docker Action
The current docker action is broken because tlmgr included within
the Pandoc image is out of date, preventing the dependencies from
being installed, which then prevents things from building.

This attempts to correct this through several fixes, which include:
- Updating the version of the Pandoc image used, which also provides
  support for rowspans & colspans in the Latex writer and fixes an
  issue in the Docx writer
- Adding an explicit step to update tlmgr in the Docker action
- Adds a workflow to run a basic check that the Dockerfile builds

This issue is because we build the Docker image on-the-fly, rather
than committing it (e.g. to Dockerhub or GitHub's Docker repository),
which is its own thing to resolve.