Attempt to fix unit tests
Despite the tests running and passing fine when under my branch
at sleevi/build-guidelines-action, once committed to cabforum, the
Link Linter test fails due to the order of "set +x".

This revealed two issues:
1. Non-determinism of the Docker output due to buffering
2. The GitHub commands from the unit test being interpreted in the
   context of the image-under-test.

This addresses the former by using a wrapper entrypoint, wrapping
in a helper script that disables input/output/stderr buffering,
and the latter by wrapping the unit tests with a prefix that indicates
the test target.