Change headings to GitHub style headings
Pandoc's default style for headings, as documented at
https://pandoc.org/MANUAL.html#headings-and-sections , is to remove
leading numbers from headings. This strips out leading section
identifiers (e.g. #foo instead of #1-foo). Pandoc does this to be
compatible with HTML 4.

This changes the Markdown reader to use the GitHub style of
creating section headers, which covers spaces to dashes, lowercases
everything, and removes punctuation other than - and _. Importantly,
this preserves numbers: "11.1.1 Foo" becomes "#11111-foo" instead of
just "#foo".

This makes it much easier to understand why a link was bad, and so
the Link checking filter is also updated to dump all the valid
identifiers whenever there is a broken link, which hopefully makes
it easier to correct the link in the event of a typo.